play lista

+++ database/migrations/2018_06_07_121015_create_parties_table.php
@@ -17,7 +17,7 @@ class CreatePartiesTable extends Migration
             $table->increments('id');
             $table->string('name', 20);
             $table->date('date');
-            $table->integer('duration');
+            $table->float('duration', 11, 2);
             $table->integer('capacity');
             $table->text('description')->nullable();
             $table->string('cover_photo', 100);
